refactor: use local import
