+++ openeogeotrellis/integrations/s3/__init__.py
@@ -1,2 +1,3 @@
 from openeogeotrellis.integrations.s3.bucket_details import BucketDetails, is_workspace_bucket
 from openeogeotrellis.integrations.s3.presigned_url import create_presigned_url
+from openeogeotrellis.integrations.s3.client import S3ClientBuilder
