+++ openeogeotrellis/integrations/s3/providers.py
