--- openeogeotrellis/integrations/s3_client/__init__.py
@@ -1,1 +1,0 @@
-from openeogeotrellis.integrations.s3_client.client import S3ClientBuilder
